Rename test/unit/airtap.js to test/unit/cli.js
