+++ test/unit/cli.js
